+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 0.0.1 `(06/17/17)`
+### 0.0.1 `(06/16/17)`
 - UX rewrite.
     - Cards with collapsible sections.
     - Card menu (e.g., delete).
@@ -11,3 +11,8 @@
 - Experimental Project boards with custom view (query).
     - Not yet editable (i.e., fixed #fav label demo query).
 - Updated user/config/testing site pages.
+
+### 0.0.2 `(06/17/17)`
+- Minimal Loggly logging (server).
+- Minimal home activity (show groups and projects).
+